Move energy comparisons to utils
